[ADD]match_usersアクセス制限

+++ app/Http/Controllers/MatchUsersController.php
@@ -11,13 +11,15 @@ class MatchUsersController extends Controller
     
     public function index()
     {
-        // 出会った方一覧をidの降順で取得
-        $match_users = MatchUser::orderBy('id', 'desc')->paginate(10);
-
-        // ユーザ一覧ビューでそれを表示
-        return view('match_users.index', [
-            'match_users' => $match_users,
-        ]);
+        if (\Auth::check()) {
+            // 出会った方一覧をidの降順で取得
+            $match_users = MatchUser::orderBy('id', 'desc')->paginate(10);
+    
+            // ユーザ一覧ビューでそれを表示
+            return view('match_users.index', [
+                'match_users' => $match_users,
+            ]);
+        }
     }
 
     public function create()
@@ -69,15 +71,21 @@ class MatchUsersController extends Controller
         // idを検索して取得
         $match_user = MatchUser::findOrFail($id);
         
+        if (\Auth::id() === $match_user->user_id) {
         return view('match_users.edit', [
             'match_user' => $match_user,
         ]);
+        }
+         // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 
     public function update(Request $request, $id)
     {
         // idを検索して取得
         $match_user = MatchUser::findOrFail($id);
+        if (\Auth::id() === $match_user->user_id) {
         // 更新
         $match_user->name = $request->name;
         $match_user->address = $request->address;
@@ -91,16 +99,27 @@ class MatchUsersController extends Controller
         $match_user->save();
         // 詳細ビューでそれを表示
         return redirect()->route('match_users.show', $match_user->id);
+        }
+         // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 
     public function destroy($id)
     {
         // idを検索して取得
         $match_user = MatchUser::findOrFail($id);
-        // 削除
-        $match_user->delete();
-        // 一覧ページ
-        return redirect()->route('match_users.index');
+        
+         if (\Auth::id() === $match_user->user_id) {
+            // 削除
+            $match_user->delete();
+            // 一覧ページ
+            return redirect()->route('match_users.index');
+         }
+         
+         // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 
 }
